+++ ltlcross_wrapper/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2019, 2020 Fanda Blahoudek
 
-__version__ = "0.7.1"
+__version__ = "0.7.1dev"
 
 from ltlcross_wrapper.analyzer import ResAnalyzer, update_resfile, merge_resfiles, remove_tools, gather_cumulative, gather_mins
 from ltlcross_wrapper.modular import Modulizer